test: address review feedback on private ingress e2e
- Replace upfront API version check with try-and-inspect pattern
  using isAPINotDeployedError(), matching kms_key_rotation.go
- Move deploySampleApp to framework package for reusability
- Remove cleanupSampleApp (cluster deletion cleans up the namespace)
- Fix missing trailing newlines
